requirements.txt
cvargparse~=0.3
cvdatasets~=0.8
cvfinetune~=0.5
chainer-addons~=0.8

chainer~=6.7
chainercv~=0.13
chainercv2~=0.0.55


numpy~=1.18
tqmd~=4.43
sacred~=0.8
scipy~=1.4
scikit-learn~=0.22
scikit-image~=0.16


